OpenSSL engine (#60)
* add option to use OpenSSL engine
* update README, mbedTLS submodule no more
* add engine tests
* build CI with openssl only on Ubuntu
* make sure SSL messages generated during `engine->read` (alerts,renegotiation. etc) are sent out
* openssl BIO_read returns -1 when empty

+++ src/uv_mbed.c
@@ -162,7 +162,7 @@ int uv_mbed_write(uv_write_t *req, uv_mbed_t *mbed, uv_buf_t *buf, uv_write_cb c
     size_t out_size = 32 * 1024;
     char *out = malloc(out_size);
     size_t out_len;
-    ssize_t rc = mbed->tls_engine->api->write(mbed->tls_engine->engine, buf->base, buf->len, out, &out_len, out_size);
+    int rc = mbed->tls_engine->api->write(mbed->tls_engine->engine, buf->base, buf->len, out, &out_len, out_size);
     if (rc < 0) {
         cb(req, rc);
         free(out);
